requirements.txt
chess
numpy
PyQt5
stopwatch.py
